server/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
google-genai
pydantic
pydantic-settings
python-multipart
pytest
